+++ src/sdhlt/sdHLRAD/transfers.cpp
@@ -146,15 +146,15 @@ bool            readtransfers(const char* const transferfile, const long numpatc
             {
 				if(g_rgb_transfers)
 				{
-                    patch->tRGBData = (rgb_transfer_data_t*)AllocBlock(patch->iData * vector_size[g_rgbtransfer_compress_type] + unused_size);
+                    patch->tRGBData = (rgb_transfer_data_t*)AllocBlock((unsigned long)(patch->iData * vector_size[g_rgbtransfer_compress_type] + unused_size));
                     hlassume(patch->tRGBData != NULL, assume_NoMemory);
-                    amtread = fread(patch->tRGBData, vector_size[g_rgbtransfer_compress_type], patch->iData, file);		    
+                    amtread = fread(patch->tRGBData, vector_size[g_rgbtransfer_compress_type], patch->iData, file);
 				}
 				else
 				{
-                    patch->tData = (transfer_data_t*)AllocBlock(patch->iData * float_size[g_transfer_compress_type] + unused_size);
+                    patch->tData = (transfer_data_t*)AllocBlock((unsigned long)(patch->iData * float_size[g_transfer_compress_type] + unused_size));
                     hlassume(patch->tData != NULL, assume_NoMemory);
-                    amtread = fread(patch->tData, float_size[g_transfer_compress_type], patch->iData, file);		    
+                    amtread = fread(patch->tData, float_size[g_transfer_compress_type], patch->iData, file);
 				}
                 if (amtread != patch->iData)
                 {
